app-engine/GCP-App-Engine-standard/python-bigquery-memcache/assignment-solution/requirements.txt
google-cloud-bigquery==2.34.4
Flask==2.1.0
appengine-python-standard